+++ doc/src/changelog.md
@@ -20,6 +20,7 @@ Version History
     in VS2013.
 -   Bugfix in rtcSetNewGeometry. Vertex buffers did not get properly
     overallocated.
+-   Replaced ImageMagick with OpenImageIO in the tutorials.
 
 ### New Features in Embree 3.0.0
 -   Switched to a new version of the API which provides improved